requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
# astro-run-dag # This package is needed for the astro run command. It will be removed before a deploy

astronomer-cosmos==1.9.2
apache-airflow-providers-postgres
